requirements.txt
google-generativeai>=0.8.3
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
requests
boto3
